test: cover issue fit finder behavior
